+++ CHANGELOG.md
@@ -1,6 +1,21 @@
+# 4.4.1
+2016-07-20
+
 ## Features
-- Get Credit Info for a Product
+- Changed `Merchant Payments` URL's
+
+# 4.4.0
+2016-07-15
+
+## Features
+- Add gateway calls for *Merchant Payments* (125524631)
 - Added `getApplicationCreditInfo` call on `ApplicationGateway` (126121685)
+
+# 4.3.0
+2016-07-06
+
+## Features
+- Get `Credit Info` call for a Product
 
 # 4.2.1
 2015-11-18